Create 238_Product_of_Array_Except_Self.c
The requirement: we cannot use divide function and need to resolve it inconstant space.
If we have the left [0,i-1] product and right[i+1,num-1] product we could easily get the "res" integer array. res[i] = left[i-1]*right[i+1]
To optimize, making save right product as a single integer  "right" and updating it while we traverse the list.
Careful for the edge situation! num=1/ i=0