Create 277_Find_the_Celebrity.c
(1) 第一个循环：
Let's say i = k, you first time switch candidate, then 0 ~ k -1 will not have candidate. The definition of a celebrity is that all the other n - 1 people know him/her but he/she does not know any of them. So if the pre candidate(0) doesn't know 1 ~ k-1, then 1 ~ k - 1 are not celebrity. As you just switch candidate, then 0 is not.
（2）check if the candidate is really satisefied.