Create 73_Set_Matrix_Zeroes.c
Store states of each row in the first of that row,
Store states of each column in the first of that column. 
Because the state of row0 and the state of column0 would occupy the same cell,  "col0" for column0. In the first phase, use matrix elements to set states in a top-down way. In the second phase, use states to set matrix elements in a bottom-up way