Create 160_Intersection_of_Two_Linked Lists.c
Smart! Construct a equivalent list based on Two sperate List
once listA hit end move toward listB, and vice versa.
Once both of them reaches the "END" it will came out of the while loop and with NULL pointer(A).
Or it will reach the conjunction together, and go ahead return it.