Create 442_Find_All_Duplicates_in_an_Array.c
I was thinking we need a way to "hash" the same number together without using extra space. I am inspired by the range of the input value (1 - N), which seems fit the index of the input array (0 - N-1).
If I hash num[i] to index num[i]-1, then I need to store it at index num[i]-1 smartly. I shouldn't make irrecoverable change to the element at that position, because I may need the element in the future. Again using the given range of input is (1 - N), if flip it to the opposite I can still recover it using Math.abs